services/stt-provider/pyproject.toml
[project]
name = "stt-provider"
version = "0.1.0"
description = "Speech-to-Text Provider Service using Deepgram"
authors = [
    {name = "Engineering Team", email = "tomdakan@gmail.com"},
]
dependencies = [
    "nats-py>=2.6.0",
    "python-dotenv>=1.0.0",
    "asyncio>=3.4.3",
    "deepgram-sdk>=5.3.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}

[project.scripts]
stt-provider = "stt_provider.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
extend = "../../ruff.toml"
